+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/DatePeriodConverters.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.staccato.presentation.converter
+package com.woowacourse.staccato.presentation
 
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
