+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/common/ToolbarHandler.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.staccato.presentation
+package com.woowacourse.staccato.presentation.common
 
 interface ToolbarHandler {
     fun onUpdateClicked()
