+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/mapper/TimelineMapper.kt
@@ -2,7 +2,7 @@ package com.woowacourse.staccato.presentation.mapper
 
 import com.woowacourse.staccato.domain.model.Timeline
 import com.woowacourse.staccato.domain.model.Travel
-import com.woowacourse.staccato.presentation.converter.convertLocalDateToDatePeriodString
+import com.woowacourse.staccato.presentation.convertLocalDateToDatePeriodString
 import com.woowacourse.staccato.presentation.timeline.model.TimelineTravelUiModel
 
 fun Timeline.toTimelineTravelUiModel(): List<TimelineTravelUiModel> {
