+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/travel/adapter/MatesViewHolder.kt
@@ -2,7 +2,7 @@ package com.woowacourse.staccato.presentation.travel.adapter
 
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import com.woowacourse.staccato.databinding.ItemMatesBinding
-import com.woowacourse.staccato.presentation.MemberUiModel
+import com.woowacourse.staccato.presentation.common.MemberUiModel
 
 class MatesViewHolder(
     private val binding: ItemMatesBinding,
