+++ android/Staccato_AN/app/src/main/res/layout/fragment_travel.xml
@@ -12,7 +12,7 @@
 
         <variable
             name="toolbarHandler"
-            type="com.woowacourse.staccato.presentation.ToolbarHandler" />
+            type="com.woowacourse.staccato.presentation.common.ToolbarHandler" />
 
         <import type="java.time.LocalDate" />
     </data>
