+++ android/Staccato_AN/app/src/main/res/layout/fragment_visit.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="toolbarHandler"
-            type="com.woowacourse.staccato.presentation.ToolbarHandler" />
+            type="com.woowacourse.staccato.presentation.common.ToolbarHandler" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
