+++ android/Staccato_AN/app/src/main/res/layout/item_mates.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="mate"
-            type="com.woowacourse.staccato.presentation.MemberUiModel" />
+            type="com.woowacourse.staccato.presentation.common.MemberUiModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
